pyproject.toml
[tool.poetry]
name = "rishut"
version = "0.1.0"
description = ""
authors = ["Roman Pavlyuchenko <romanpavlyuchenko@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
flake8 = "^6.0.0"
python-dotenv = "^0.21.1"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
